Перенести логику из классов
