+++ src/test/java/ru/yandex/practicum/filmorate/model/UserTest.java
@@ -3,6 +3,7 @@ package ru.yandex.practicum.filmorate.model;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import ru.yandex.practicum.filmorate.controller.UserController;
 
 import java.time.LocalDate;
 
@@ -10,7 +11,7 @@ class UserTest {
 
     @BeforeEach
     void setUp() {
-        User.users.clear();
+        UserController.users.clear();
     }
 
     @Test
@@ -22,10 +23,10 @@ class UserTest {
             user.setLogin("Login");
             user.setName("Name");
             user.setBirthday(LocalDate.of(2000, 6, 11));
-            User.addUser(user);
+            UserController.add(user);
         }
 
-        Assertions.assertEquals(creatingTimes, User.getUsersList().size());
+        Assertions.assertEquals(creatingTimes, UserController.getUsers().size());
     }
 
     @Test
@@ -38,7 +39,7 @@ class UserTest {
             user.setLogin("Login");
             user.setName("Name");
             user.setBirthday(LocalDate.of(2000, 6, 11));
-            lastAddedUser = User.addUser(user);
+            lastAddedUser = UserController.add(user);
         }
 
         Assertions.assertEquals(creatingTimes, lastAddedUser.getId());
